backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
dateparser==1.2.0
distro==1.9.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
MarkupSafe==3.0.1
multidict==6.1.0
newscatcherapi==0.7.3
numpy==2.1.2
openai==1.51.2
packaging==24.1
pinecone==5.3.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.14.0
